### 2023-04-18 #5383 @editorialbot
[PRE REVIEW]: NORDic: a Network-Oriented package for the Repurposing of Drugs
**Submitting author:** @clreda (<a href="http://orcid.org/0000-0003-3238-0258">Clémence Réda</a>)
**Repository:** https://github.com/clreda/NORDic
**Branch with paper.md** (empty if default branch):
**Version:** v2.4.2
