+++ paper/paper.bib
@@ -6,6 +6,7 @@
   number={4},
   pages={1798--1808},
   year={2013},
+  doi={10.1172/JCI66220},
   publisher={Am Soc Clin Investig}
 }
 
@@ -17,6 +18,7 @@
   number={10},
   pages={770--780},
   year={2008},
+  doi={10.1038/nrm2503},
   publisher={Nature Publishing Group UK London}
 }
 
@@ -28,6 +30,7 @@
   number={3},
   pages={563--585},
   year={1973},
+  doi={10.1016/0022-5193(74)90172-6},
   publisher={Elsevier}
 }
 
@@ -38,6 +41,7 @@
   volume={22},
   number={3},
   pages={437--467},
+  doi={10.1016/0022-5193(69)90015-0},
   year={1969},
   publisher={Elsevier}
 }
@@ -47,6 +51,7 @@
   author={Moon, Kyungduk and Lee, Kangbok and Paulev{\'e}, Lo{\"\i}c},
   journal={arXiv preprint arXiv:2212.12756},
   year={2022}
+  doi={10.48550/arXiv.2212.12756},
 }
 
 @article{Zerrouk2020,
@@ -57,18 +62,8 @@
   number={1},
   pages={16236},
   year={2020},
+  doi={10.1038/s41598-020-73147-4},
   publisher={Nature Publishing Group UK London}
-}
-
-@article{Stoll2017,
-  title={MaBoSS 2.0: an environment for stochastic Boolean modeling},
-  author={Stoll, Gautier and Caron, Barth{\'e}l{\'e}my and Viara, Eric and Dugourd, Aur{\'e}lien and Zinovyev, Andrei and Naldi, Aur{\'e}lien and Kroemer, Guido and Barillot, Emmanuel and Calzone, Laurence},
-  journal={Bioinformatics},
-  volume={33},
-  number={14},
-  pages={2226--2228},
-  year={2017},
-  publisher={Oxford University Press}
 }
 
 @article{Srivastava2018,
@@ -79,6 +74,7 @@
   number={1},
   pages={3561},
   year={2018},
+  doi={10.1038/s41467-018-06008-4},
   publisher={Nature Publishing Group UK London}
 }
 
@@ -90,6 +86,7 @@
   number={1},
   pages={1--12},
   year={2016},
+  doi={10.1038/npjsba.2016.15},
   publisher={Nature Publishing Group}
 }
 
@@ -101,6 +98,7 @@
   number={6},
   pages={e1391--e1396},
   year={2021},
+  doi={10.1093/cid/ciab350},
   publisher={Oxford University Press US}
 }
 
@@ -112,6 +110,7 @@
   number={23},
   pages={5792--5799},
   year={2017},
+  doi={10.1073/pnas.1610622114},
   publisher={National Acad Sciences}
 }
 
@@ -123,17 +122,8 @@
   number={2},
   pages={216--226},
   year={2008},
+  doi={10.1016/j.cell.2008.09.050},
   publisher={Elsevier}
-}
-
-@article{Montagud2022,
-  title={Patient-specific Boolean models of signalling networks guide personalised treatments},
-  author={Montagud, Arnau and B{\'e}al, Jonas and Tobalina, Luis and Traynard, Pauline and Subramanian, Vigneshwari and Szalai, Bence and Alf{\"o}ldi, R{\'o}bert and Pusk{\'a}s, L{\'a}szl{\'o} and Valencia, Alfonso and Barillot, Emmanuel and others},
-  journal={Elife},
-  volume={11},
-  pages={e72626},
-  year={2022},
-  publisher={eLife Sciences Publications Limited}
 }
 
 @article{Bolouri2003,
@@ -144,6 +134,7 @@
   number={16},
   pages={9371--9376},
[…]
