+++ paper/paper.bib
@@ -50,7 +50,7 @@
   title={Computational Complexity of Minimal Trap Spaces in Boolean Networks},
   author={Moon, Kyungduk and Lee, Kangbok and Paulev{\'e}, Lo{\"\i}c},
   journal={arXiv preprint arXiv:2212.12756},
-  year={2022}
+  year={2022},
   doi={10.48550/arXiv.2212.12756},
 }
 
@@ -206,7 +206,7 @@
   booktitle={2019 IEEE 31st International Conference on Tools with Artificial Intelligence (ICTAI)},
   pages={34--41},
   year={2019},
-  organization={IEEE}
+  organization={IEEE},
   doi={10.1109/ictai.2019.00014},
 }
 
@@ -242,7 +242,7 @@
   number={10},
   pages={1378--1380},
   year={2010},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
   doi={10.1093/bioinformatics/btq124},
 }
 
@@ -276,7 +276,7 @@
   booktitle={International Conference on Artificial Intelligence and Statistics},
   pages={1108--1116},
   year={2021},
-  organization={PMLR}
+  organization={PMLR},
   doi={10.48550/arXiv.2103.10070},
 }
 
@@ -364,7 +364,7 @@
   number={6},
   pages={1437--1452},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi={10.1016/j.cell.2017.10.049},
 }
 
@@ -398,7 +398,7 @@
   number={D1},
   pages={D605--D612},
   year={2021},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
   doi={10.1093/nar/gkaa1074},
 }
 
