### 2023-06-22 f759871
Update functions.py
Fix Python3.10 issue with reading NaNs
