### 2023-07-10 #5532 @cauliyang
# My Current Comments

## Installation

- `pip install NORDic` failed on macOS arm64
```
ERROR: Failed building wheel for pyeda
```
- `conda install -c colomoto -y -q maboss` failed on macOS arm64
```
The following package could not be installed
└─ maboss does not exist (perhaps a typo or a missing channel).
```

-  `conda install -c creda -y -q nordic` failed on macOS arm64.
```
The following package could not be installed
└─ nordic is not installable because it requires
   └─ bonesis >=0.4.91 , which does not exist (perhaps a missing channel).
```
