### 2023-07-10 2cb6f15
add platforms to README
