### 2023-07-10 #5532 @clreda
Hello @cauliyang, thanks for reviewing this package!

Could you please detail the error in building the wheel when using pip? Did you test that first, or did you try installing maboss first?

Noted for the platforms, as I have mainly tried (and made people try) on Linux. I will correct meta.yaml while waiting for your reply.
