### 2023-07-14 #5532 @cauliyang
Hi @clreda , I figured it out finally, the problem is thanks to the package `pyeda`.
The package `NORDic` can be installed via `pip` and `conda` successfully. But the readme may be confusing.  we do not need to install dependencies if we install via `conda`. It is better to clarify when we need to install dependencies by ourselves in the installation information.

# Reproducibility

it seems like the file you use in the notebook does not exist. for example `ToyOndine/inferred_maximal_solution.png` in `NORDic Network Simulations.ipynb`

# Test

The test looks a little bit messy, and it is better to use `pytest` or `unittest` to manage all your tests.

# Others

- use a formatter (e.g black) can make code more clean and readable
- follow https://peps.python.org/pep-0008/ or https://google.github.io/styleguide/pyguide.html will be better
