### 2023-07-18 05ba52e
modify readme in pypi/conda; clarify order of running notebooks
