### 2023-07-18 #5532 @RECeSS-EU-Project
@cauliyang I am glad to hear that the problem is solved! I have corrected the README to include the differences between the installation with PyPI and Conda.

About your reproducibility concern: the notebooks are supposed to be run in a given order. Indeed, most of the notebooks rely on the network inferred in notebook ``NORDic CoLoMoTo.ipynb``, as the inference code might be a bit too large to add to every notebook. I will clarify this dependency in all notebooks.

I will reformat the testing code to use ``unittest`` as suggested. I will check out formatter Black and the guidelines you have mentioned, thank you!
