### 2023-08-12 #5532 @sevgisu0227
Installation worked fine, I used pip install option and manually installed the dependencies both on Mac OS and Ubuntu 20.04 LTS. One recommendation would be to include recommended minimum system  requirements.

**Functionality documentation**

I am non academic user, so this may be specific to me, I had to ask someone else for lincs sign-up. providing an alternative, ie. downloading the files from clue which does not require sign-up and processing it could be more helpful. For reproducibility and version control purposes, I would provide these databases as reference files and avoid API access all together.

providing a summary of recommended (or potentially default) parameters, adding usage function for the parameters and short descriptions would make it a lot more user friendly.
