+++ tests/tests_NI.py
@@ -9,72 +9,78 @@ from subprocess import call
 
 ## I. Building a small regulatory model of CCHS (**NORDic NI**)
 
-## Registration to databases
-DisGeNET_credentials = "../tests/credentials_DISGENET.txt"
-STRING_credentials = "../tests/credentials_STRING.txt"
-LINCS_credentials = "../tests/credentials_LINCS.txt"
-## Parameters
-seed_number=123456
-njobs=max(1,cpu_count()-2) ## all available threads but 2
-file_folder="ToyOndine/"
-taxon_id=9606 # human species
-disease_cids=["C1275808"] ## Concept ID of Ondine syndrome
-cell_lines=["NPC", "SHSY5Y"] # brain cell lines in LINCS L1000
-## Information about the disease
-DISGENET_args = {"credentials": DisGeNET_credentials, "disease_cids": disease_cids}
-## Selection of parameters relative to the prior knowledge network 
-STRING_args = {"credentials": STRING_credentials, "score": 0}
-EDGE_args = {"tau": 0, "filter": True, "connected": True}
-accept_nonRNA=True
-preserve_network_sign=True
-## Selection of parameters relative to experimental constraints
-LINCS_args = {"path_to_lincs": "../lincs/", "credentials": LINCS_credentials, "cell_lines": cell_lines, 
+class TestNI(unittest.TestCase):
+
+    def create_instance(self):
+        ## Registration to databases
+        DisGeNET_credentials = "../tests/credentials_DISGENET.txt"
+        STRING_credentials = "../tests/credentials_STRING.txt"
+        LINCS_credentials = "../tests/credentials_LINCS.txt"
+        ## Parameters
+        seed_number=123456
+        njobs=max(1,cpu_count()-2) ## all available threads but 2
+        file_folder="ToyOndine/"
+        taxon_id=9606 # human species
+        disease_cids=["C1275808"] ## Concept ID of Ondine syndrome
+        cell_lines=["NPC", "SHSY5Y"] # brain cell lines in LINCS L1000
+        ## Information about the disease
+        DISGENET_args = {"credentials": DisGeNET_credentials, "disease_cids": disease_cids}
+        ## Selection of parameters relative to the prior knowledge network 
+        STRING_args = {"credentials": STRING_credentials, "score": 0}
+        EDGE_args = {"tau": 0, "filter": True, "connected": True}
+        accept_nonRNA=True
+        preserve_network_sign=True
+        ## Selection of parameters relative to experimental constraints
+        LINCS_args = {"path_to_lincs": "../lincs/", "credentials": LINCS_credentials, "cell_lines": cell_lines, 
               "thres_iscale": None}
-SIG_args = {"bin_thres": 0.5}
-force_experiments=False
-## Selection of parameters relative to the inference of networks
-BONESIS_args = {"limit": 1, "exact": True, "max_maxclause": 3}
-## Advanced
-DESIRABILITY = {"DS": 3, "CL": 3, "Centr": 3, "GT": 1}
+        SIG_args = {"bin_thres": 0.5}
+        force_experiments=False
+        ## Selection of parameters relative to the inference of networks
+        BONESIS_args = {"limit": 1, "exact": True, "max_maxclause": 3}
+        ## Advanced
+        DESIRABILITY = {"DS": 3, "CL": 3, "Centr": 3, "GT": 1}
+        call("mkdir -p "+file_folder, shell=True)
 
[…]
