+++ tests/tests_PKN.py
@@ -3,6 +3,8 @@
 import NORDic
 import os
 import pandas as pd
+import unittest
+from subprocess import Popen
 
 class TestPKN(unittest.TestCase):
 
@@ -34,7 +36,7 @@ class TestPKN(unittest.TestCase):
 
     def test_merge_network_PPI(self):
         ## Retrieve networks from the STRING database
-        from NORDic.UTILS.STRING_utils import merge_network_PPI
+        from NORDic.UTILS.utils_network import merge_network_PPI
         file_folder="ToyOndine/"
         core_genes = ["PHOX2B", "RET", "BDNF", "ASCL1", "EDN3", "GDNF"]
         taxon_id = 9606
@@ -47,6 +49,7 @@ class TestPKN(unittest.TestCase):
         return final_network, core_genes
 
     def test_determine_threshold(self):
+        file_folder="ToyOndine/"
         final_network, core_genes = self.test_merge_network_PPI()
         from NORDic.UTILS.utils_network import determine_edge_threshold
         threshold = determine_edge_threshold(final_network, core_genes)
@@ -68,14 +71,20 @@ class TestPKN(unittest.TestCase):
 
     def test_plot_influence_graph(self):
         from NORDic.UTILS.utils_plot import plot_influence_graph
+        file_folder="ToyOndine/"
         final_network = self.test_determine_threshold()
         plot_influence_graph(final_network, "preferredName_A", "preferredName_B", "sign", direction_col="directed", fname=file_folder+"graph_final", optional=True)
 
-    def test_plot_influence_graph(self):
+    def test_full_pipeline(self):
+        from NORDic.UTILS.STRING_utils import get_network_from_STRING
         from NORDic.UTILS.utils_network import aggregate_networks
-        file_folder="ToyOndine/"
+        file_folder="ToyOndine_full/"
+        Popen(('cp -r ToyOndine/ '+file_folder).split(" "))
         core_genes = ["PHOX2B", "RET", "BDNF", "ASCL1", "EDN3", "GDNF"]
         taxon_id = 9606
+        network = get_network_from_STRING(core_genes, taxon_id, min_score=0., network_type="functional",
+                add_nodes=0, app_name="NORDic PKN", version="11.5", quiet=0)
+        network.to_csv(file_folder+"network.csv")
         final_network2 = aggregate_networks(file_folder, core_genes, taxon_id, 0, "functional", "NORDic whole PKN", quiet=1)
         glist2 = list(set(reduce(lambda x,y: x+y, [list(final_network2[c]) for c in ["preferredName_A", "preferredName_B"]])))
         components2 = get_weakly_connected(final_network2, glist2, score_col="score")
