### 2023-08-20 #5532 @clreda
Dear @cauliyang I have rewritten the automated tests in the package using unittest. Please let me know if it is better to you

Dear @sevgisu0227 thank you for accepting to review this paper. The main issue I see with providing the database could be space-consuming for most of them (ex. LINCS and STRING, in several dozens of GB). Using the API version of the database also ensures that the query results are up to date and do not manually require me to monitor all the databases and redownload them at each update. As an answer to your reproducibility problem, I can possibly suggest writing down the date of creation of the model, which can be connected to the latest version at that date of each database.
I will add further documentation in the docstrings in the next weeks, including short descriptions, usage examples, and default/recommended values.
