+++ src/NORDic/UTILS/utils_grn.py
@@ -19,34 +19,67 @@ from NORDic.UTILS.utils_state import quantile_normalize
 
 def desirability(x, f_weight_di, A=0, B=1):
     '''
-        Harrington's desirability function, used by [1]
-        Convert a list of functions to maximize into a single scalar function to maximize with values in [@A,@B]
-        @param\tx\tPoint
-        @param\tf_weight_di\tPython dictionary: function with arguments as the same type as x, and associated weight
-        @param\tA\tPython float[default=0]: lower bound of the function interval
-        @param\tB\tPython float[default=1]: upper bound of the function interval
-        @return\tdes(x)\tvalue of the desirability function at point x
-        [1] http://ceur-ws.org/Vol-2488/paper17.pdf
+    Harrington's desirability function, used by [1]
+    Converts a list of functions to maximize into a single scalar function to maximize with values in [@A,@B]
+
+    [1] http://ceur-ws.org/Vol-2488/paper17.pdf
         https://cran.r-project.org/web/packages/desirability/vignettes/desirability.pdf
+
+    ...
+
+    Parameters
+    ----------
+    x : datapoint
+        any input to functions in f_weight_di
+    f_weight_di : Python dictionary
+        function with arguments as the same type as x, and associated weight
+    A : Python float
+        [default=0] : lower bound of the function interval
+    B : Python float
+        [default=1] : upper bound of the function interval
+
+    Returns
+    ----------
+    des(x) : Python float
+        value of the desirability function at point x
     '''
     values = [np.exp(-np.exp(-((B-A)+weight/(B-A)*funct(x)))) for funct, weight in f_weight_di.items()]
     return float(np.exp(np.mean(np.log(values))))
 
 def DS(influences):
     '''
-        Computes the number of edges over the maximum number of possible connections between the nodes in the network
-        @param\tinfluences\tPandas DataFrame: rows/[genes] x columns/[genes]
-        @return\tDS\tPython float: network density
+    Computes the number of edges over the maximum number of possible connections between the nodes in the network
+
+    ...
+
+    Parameters
+    ----------
+    influences : Pandas DataFrame
+        rows/[genes] x columns/[genes]
+
+    Returns
+    ----------
+    DS : Python float
+        network density
     '''
     n = influences.shape[0]
     return 2*np.sum(np.sum(np.triu(np.abs(influences.values)), axis=1), axis=0)/float(n*(n-1))
 
 def CL(influences):
     '''
-        Computes the average of node-wise clustering coefficients
-        The clustering coefficient of a node is the ratio of the degree of the considered node and the maximum possible number of connections such that this node and its current neighbors form a clique
-        @param\tinfluences\tPandas DataFrame: rows/[genes] x columns/[genes]
-        @return\tCL\tPython float: network clustering coefficient
+    Computes the average of node-wise clustering coefficients. The clustering coefficient of a node is the ratio of the degree of the considered node and the maximum possible number of connections such that this node and its current neighbors form a clique
+
+    ...
+
+    Parameters
+    ----------
+    influences : Pandas DataFrame
+        rows/[genes] x columns/[genes]
+
+    Returns
+    ----------
+    CL : Python float
+        network clustering coefficient
     '''
     n = influences.shape[0]
     network_outdegree = np.sum(np.abs(influences.values), axis=1)
@@ -58,9 +91,19 @@ def CL(influences):
 
 def Centr(influences):
     '''
-        Computes the network centralization, which is correlated with the similarity of the network to a graph with a star topology
[…]
