+++ src/NORDic/UTILS/LINCS_utils.py
@@ -12,9 +12,19 @@ from functools import reduce
 
 def get_user_key(fname):
     '''
-        Retrieves user key for interacting with LINCS L1000 CLUE API
-        @param\tfname\tPython character string: path to file containing credentials for LINCS L1000 (first line: username, second line: password, third line: user key)
-        @return\tuser_key\tPython character string: identifier for the LINCS L1000 CLUE API
+    Retrieves user key for interacting with LINCS L1000 CLUE API
+
+    ...
+
+    Parameters
+    ----------
+    fname : Python character string
+        path to file containing credentials for LINCS L1000 (first line: username, second line: password, third line: user key)
+
+    Returns
+    ----------
+    user_key : Python character string
+        identifier for the LINCS L1000 CLUE API
     '''
     with open(fname, "r") as f:
         user_key = f.read().split("\n")[2]
@@ -22,10 +32,20 @@ def get_user_key(fname):
 
 def convert_ctrlgenes_EntrezGene(taxon_id):
     '''
-        Retrieves EntrezID from control genes in LINCS L1000 [1]
-        @param\ttaxon_id\tPython integer: NCBI taxonomy ID
-        @return\tlincs_specific_ctl_genes\tPython character string list: list of EntrezGene IDs for all genes in input list
-        [1] doi.org/10.1002/psp4.12107 
+    Retrieves EntrezID from control genes in LINCS L1000 [1]
+    [1] doi.org/10.1002/psp4.12107 
+
+    ...
+
+    Parameters
+    ----------
+    taxon_id : Python integer
+        NCBI taxonomy ID
+
+    Returns
+    ----------
+    lincs_specific_ctl_genes : Python character string list
+        list of EntrezGene IDs for all genes in input list
     '''
     lincs_specific_ctl_genes = ["ACTB", "CHMP2A", "EEF1A1", "EMC7", "GAPDH", "GPI", "PSMB2", "PSMB4", "RAB7A", "REEP5", "SNRPD3", "TUBA1A", "VCP"]
     from NORDic.UTILS.utils_data import request_biodbnet
@@ -41,12 +61,25 @@ lincs_api_url = "https://api.clue.io/api"
 
 def build_url(endpoint, method, params, user_key=None):
     '''
-        Builds the request to CLUE API
-        @param\tendpoint\tPython character string: in ["sigs", "cells", "genes", "perts", "plates", "profiles", "rep_drugs", "rep_drug_indications", "pcls"]
-        @param\tmethod\tPython character string: in ["count", "filter", "distinct"]
-        @param\tparams\tPython dictionary: additional arguments for the request
-        @param\tuser_key\tPython character string
-        @return\turl\tPython character string: URL of request
+    Builds the request to CLUE API
+
+    ...
+
+    Parameters
+    ----------
+    endpoint : Python character string
+        in ["sigs", "cells", "genes", "perts", "plates", "profiles", "rep_drugs", "rep_drug_indications", "pcls"]
+    method : Python character string
+        in ["count", "filter", "distinct"]
+    params : Python dictionary
+        additional arguments for the request
+    user_key : Python character string
+        [default=None] : API key for LINCS CLUE.io
+
+    Returns
+    ----------
+    url : Python character string
+        URL of request
     '''
     assert user_key
     assert endpoint in ["sigs", "cells", "genes", "perts", "plates", "profiles", "rep_drugs", "rep_drug_indications", "pcls"]
@@ -81,11 +114,23 @@ def build_url(endpoint, method, params, user_key=None):
 
 def post_request(url, quiet=True, pause_time=1):
     '''
-        Post request to API
-        @param\turl\tPython character string: URL formatted as in build_url
-        @param\tquiet\tPython bool[default=True]
-        @param\tpause_time\tPython integer[default=1]: minimum time in seconds between each request
[…]
