+++ docs/conf.py
@@ -11,7 +11,7 @@ sys.path.insert(0, os.path.abspath("../src"))
 project = 'NORDic'
 copyright = '2023, Clémence Réda, Andrée Delahaye-Duriez'
 author = 'Clémence Réda, Andrée Delahaye-Duriez'
-release = '2.4.3'
+release = 'v2.4.3'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
@@ -22,7 +22,7 @@ extensions = [
 ]
 
 templates_path = ['_templates']
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+exclude_patterns = []
 
 
 
