+++ .github/workflows/gh-pages.yml
@@ -4,9 +4,11 @@ on: [push]
 #on:
 #  release:
 #    types: [published]
+permissions:
+    contents: write
 
 jobs:
-  build:
+  docs:
 
     runs-on: ubuntu-latest
     defaults:
