### 2023-08-31 #5532 @clreda
Dear @cauliyang I have found that using directly the [CoLoMoTo docker](https://github.com/colomoto/colomoto-docker) might answer the concern of users wishing to avoid environment configuration issues, as it includes the first tutorial notebook in NORDic. I have specified how to use it in the [README](https://github.com/clreda/NORDic#quick-access-to-nordic) as well as in the [documentation website](https://clreda.github.io/NORDic). This documentation website might answer your second concern, about concise API documentation. I have also tackled one of your earliest concerns about submitting to a larger conda channel. I chose ``bioconda`` (see [pull request](https://github.com/bioconda/bioconda-recipes/pull/42756)). I hope to have everything fixed in the next weeks.

Dear @sevgisu0227 I have added further documentation in the docstrings, with short descriptions and default/recommended values (please refer to the [documentation website](https://clreda.github.io/NORDic)). I haven't added the usage examples, because since most of the arguments in functions rely on custom structures introduced in NORDic, showcasing NORDic functionalities through notebooks might make those functions easier to use that long docstring descriptions.

Please let me know if you have other concerns.
