+++ .github/workflows/gh-pages.yml
@@ -1,11 +1,8 @@
 name: NORDic documentation
 
-on: [push]
-#on:
-#  release:
-#    types: [published]
-permissions:
-    contents: write
+on:
+  release:
+    types: [published]
 
 jobs:
   docs:
@@ -24,7 +21,7 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
-      - name: Install dependencies and build stanscofi
+      - name: Install dependencies and build NORDic
         run: |
           python -m pip install --upgrade pip wheel
           python -m pip install -r pip/requirements.txt
