+++ docs/modules.rst
@@ -4,4 +4,8 @@ NORDic
 .. toctree::
    :maxdepth: 4
 
-   NORDic
+   NORDic_NI
+   NORDic_PMR
+   NORDic_DS
+   NORDic_DR
+   NORDic_UTILS
