+++ docs/NORDic_DR.rst
@@ -1,0 +1,26 @@
+NORDic_DR module
+=================
+
+NORDic.NORDic\_DR.bandits module
+----------------------------------
+
+.. automodule:: NORDic.NORDic_DR.bandits
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+NORDic.NORDic\_DR.functions module
+----------------------------------
+
+.. automodule:: NORDic.NORDic_DR.functions
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+NORDic.NORDic\_DR.utils module
+----------------------------------
+
+.. automodule:: NORDic.NORDic_DR.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
