+++ docs/NORDic_NI.rst
@@ -1,0 +1,10 @@
+NORDic NI module
+=================
+
+NORDic.NORDic\_NI.functions module
+--------------------------------------------------
+
+.. automodule:: NORDic.NORDic_NI.functions
+   :members:
+   :undoc-members:
+   :show-inheritance:
