+++ docs/NORDic_UTILS.rst
@@ -1,72 +1,5 @@
-NORDic package
-=================
-
-Submodules
-----------
-
-NORDic.NORDic\_NI.functions module
---------------------------------------------------
-
-.. automodule:: NORDic.NORDic_NI.functions
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.NORDic\_PMR.functions module
---------------------------------------------------
-
-.. automodule:: NORDic.NORDic_PMR.functions
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.NORDic\_DS.functions module
--------------------------------------------------------
-
-.. automodule:: NORDic.NORDic_DS.functions
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.NORDic\_DS.get_drug_signatures module
----------------------------------------------------------------
-
-.. automodule:: NORDic.NORDic_DS.get_drug_signatures
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.NORDic\_DS.get_drug_targets module
----------------------------------------------------------------
-
-.. automodule:: NORDic.NORDic_DS.get_drug_targets
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.NORDic\_DR.bandits module
-----------------------------------
-
-.. automodule:: NORDic.NORDic_DR.bandits
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.NORDic\_DR.functions module
-----------------------------------
-
-.. automodule:: NORDic.NORDic_DR.functions
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.NORDic\_DR.utils module
-----------------------------------
-
-.. automodule:: NORDic.NORDic_DR.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
+NORDic UTILS module
+====================
 
 NORDic.UTILS.DISGENET_utils module
 ---------------------------------------
