+++ _sources/NORDic.rst.txt
@@ -1,0 +1,22 @@
+NORDic package
+==============
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   NORDic.NORDic_DR
+   NORDic.NORDic_DS
+   NORDic.NORDic_NI
+   NORDic.NORDic_PMR
+   NORDic.UTILS
+
+Module contents
+---------------
+
+.. automodule:: NORDic
+   :members:
+   :undoc-members:
+   :show-inheritance:
