+++ _sources/NORDic_PMR.rst.txt
@@ -1,0 +1,10 @@
+NORDic_PMR module
+=================
+
+NORDic.NORDic\_PMR.functions module
+--------------------------------------------------
+
+.. automodule:: NORDic.NORDic_PMR.functions
+   :members:
+   :undoc-members:
+   :show-inheritance:
