+++ _sources/NORDic_NI.rst.txt
@@ -1,0 +1,10 @@
+NORDic_NI module
+=================
+
+NORDic.NORDic\_NI.functions module
+--------------------------------------------------
+
+.. automodule:: NORDic.NORDic_NI.functions
+   :members:
+   :undoc-members:
+   :show-inheritance:
