+++ _sources/NORDic_UTILS.rst.txt
@@ -1,82 +1,0 @@
-NORDic_UTILS module
-====================
-
-NORDic.UTILS.DISGENET_utils module
----------------------------------------
-
-.. automodule:: NORDic.UTILS.DISGENET_utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.UTILS.LINCS_utils module
----------------------------------
-
-.. automodule:: NORDic.UTILS.LINCS_utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.UTILS.STRING_utils module
----------------------------------
-
-.. automodule:: NORDic.UTILS.STRING_utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.UTILS.utils_data module
----------------------------------
-
-.. automodule:: NORDic.UTILS.utils_data
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.UTILS.utils_exp module
----------------------------------
-
-.. automodule:: NORDic.UTILS.utils_exp
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.UTILS.utils_grn module
----------------------------------
-
-.. automodule:: NORDic.UTILS.utils_grn
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.UTILS.utils_network module
----------------------------------
-
-.. automodule:: NORDic.UTILS.utils_network
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.UTILS.utils_plot module
----------------------------------
-
-.. automodule:: NORDic.UTILS.utils_plot
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.UTILS.utils_sim module
----------------------------------
-
-.. automodule:: NORDic.UTILS.utils_sim
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.UTILS.utils_state module
----------------------------------
-
-.. automodule:: NORDic.UTILS.utils_state
-   :members:
-   :undoc-members:
-   :show-inheritance:
