+++ _sources/NORDic.NORDic_DR.rst.txt
@@ -1,37 +1,0 @@
-NORDic.NORDic\_DR package
-=========================
-
-Submodules
-----------
-
-NORDic.NORDic\_DR.bandits module
---------------------------------
-
-.. automodule:: NORDic.NORDic_DR.bandits
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.NORDic\_DR.functions module
-----------------------------------
-
-.. automodule:: NORDic.NORDic_DR.functions
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.NORDic\_DR.utils module
-------------------------------
-
-.. automodule:: NORDic.NORDic_DR.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: NORDic.NORDic_DR
-   :members:
-   :undoc-members:
-   :show-inheritance:
