+++ _sources/NORDic.UTILS.rst.txt
@@ -1,93 +1,0 @@
-NORDic.UTILS package
-====================
-
-Submodules
-----------
-
-NORDic.UTILS.DISGENET\_utils module
------------------------------------
-
-.. automodule:: NORDic.UTILS.DISGENET_utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.UTILS.LINCS\_utils module
---------------------------------
-
-.. automodule:: NORDic.UTILS.LINCS_utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.UTILS.STRING\_utils module
----------------------------------
-
-.. automodule:: NORDic.UTILS.STRING_utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.UTILS.utils\_data module
--------------------------------
-
-.. automodule:: NORDic.UTILS.utils_data
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.UTILS.utils\_exp module
-------------------------------
-
-.. automodule:: NORDic.UTILS.utils_exp
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.UTILS.utils\_grn module
-------------------------------
-
-.. automodule:: NORDic.UTILS.utils_grn
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.UTILS.utils\_network module
-----------------------------------
-
-.. automodule:: NORDic.UTILS.utils_network
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.UTILS.utils\_plot module
--------------------------------
-
-.. automodule:: NORDic.UTILS.utils_plot
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.UTILS.utils\_sim module
-------------------------------
-
-.. automodule:: NORDic.UTILS.utils_sim
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.UTILS.utils\_state module
---------------------------------
-
-.. automodule:: NORDic.UTILS.utils_state
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: NORDic.UTILS
-   :members:
-   :undoc-members:
-   :show-inheritance:
