+++ _sources/NORDic_DS.rst.txt
@@ -1,26 +1,0 @@
-NORDic_DS module
-=================
-
-NORDic.NORDic\_DS.functions module
--------------------------------------------------------
-
-.. automodule:: NORDic.NORDic_DS.functions
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.NORDic\_DS.get_drug_signatures module
----------------------------------------------------------------
-
-.. automodule:: NORDic.NORDic_DS.get_drug_signatures
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-NORDic.NORDic\_DS.get_drug_targets module
----------------------------------------------------------------
-
-.. automodule:: NORDic.NORDic_DS.get_drug_targets
-   :members:
-   :undoc-members:
-   :show-inheritance:
