+++ _sources/modules.rst.txt
@@ -1,0 +1,11 @@
+NORDic
+======
+
+.. toctree::
+   :maxdepth: 4
+
+   NORDic_NI
+   NORDic_PMR
+   NORDic_DS
+   NORDic_DR
+   NORDic_UTILS
