### 2023-09-03 #5532 @cauliyang
Hi @clreda, Thanks for your effort and consideration! A Docker environment is also a better alternative. Bioconda is a wise choice and promotes the tool to the larger community. The document systems are on the right track as well. In summary, your changes resolve my concerns. Keep the library alive and friendly!
