### 2023-09-03 #5532 @clreda
Dear @cauliyang thanks a lot for your feedback, which allows NORDic to be more accessible and useful to the community!
