### 2023-09-13 #5532 @RECeSS-EU-Project
Dear @sevgisu0227 thank you very much for your feedback that improved the user experience with NORDic!
