### 2023-09-13 #5532 @clreda
Dear @sevgisu0227 thank you very much for your feedback that improved the user experience with NORDic!
