### 2023-09-18 #5532 @clreda
The latest release version on GitHub is v2.4.4. The DOI of the corresponding Zenodo release is 10.5281/zenodo.8355529.
