### 2023-09-28 #5532 @lpantano
Thank you! I have been reading the paper again, and I think we should reduce the information on it to be around 1500 words, no counting references. Right now the text contains more than 2000 words (can you check that?). If that is true, I am thinking about the following:
1-reduce introduction to half a page.
2-For these sections: Automated identification of disease-related Boolean networks and Prioritization of master regulators in Boolean networks, there first couple of paragraphs is about the current work and issues to solve. I think this is very important and I will keep this information in the documentation for instance, but maybe in the paper, it would be good to focus last part of the section that explain NORDic implementation and try to summarize in one paragraph the issues.

Maybe the current information can be in the docs (if they are not yet there), and the paper can be shorter. I know it is a lot of work to explain the current status of the field and the good things NORDic is bringing.

Let me know if you think is possible to bring the text to ~1500 words. Thanks in advance.
