+++ paper/paper.md
@@ -27,33 +27,21 @@ bibliography: paper.bib
 
 # Introduction
 
-Genes, proteins and messenger RNAs are shown to interact on each other in order to modulate gene activity. Conversely, gene activity impacts
-protein production, and consequently triggers the chemical reactions needed for survival in healthy individuals. As such, perturbations
-of these gene regulatory interactions, through (epi)genetic and/or environmental factors, might cause diseases: e.g., the suppression of the 
-activity of gene SCN1A is linked to a specific type of epilepsy called Dravet syndrome, both in mice and humans [@Kalume2013]. Gene 
+Genes, proteins and messenger RNAs are shown to interact on each other in order to modulate gene activity. Gene 
 regulatory networks, which are graphs connecting biological entities according to their known regulatory interactions, are useful models 
 that enable a better understanding of those regulatory mechanisms [@Karlebach2008]. 
 
 In particular, one type of gene regulatory networks, called Boolean networks, allows the definition of so-called regulatory functions 
 [@Thomas1973; @Kauffman1969]. Those functions are specific to each node in the graph, and determine the activity of this node 
 according to its regulators. Those functions are defined on the Boolean domain (**True** or **False**), meaning that we only consider binary 
-gene activities. Subsequently, studying this type of networks as a dynamical system (and determining their basins of attraction, for instance) 
-remains rather tractable [@Moon2022]. The potential applications are numerous. Taking into account network dynamics should improve tools originally developed using non Boolean networks, such as the Transcription Factor Influence score in CoRegNet [@Nicolle2015] for the identification of interesting 
-biomarkers, or drug repurposing. For the latter, a library of druggable molecules can be screened for good candidates based on the paradigm of "signature reversion"; a good drug candidate should be able to "reverse" the gene activity profile 
-associated with a diseased individual [@Duan2016; @Delahaye2016; @Musa2018]. That is, such a drug would stimulate abnormally weakly activated genes with 
-respect to healthy individuals, and vice-versa.
-This screening approach may considerably speed up drug development, especially for rare or tropical neglected diseases [@Walker2021].
+gene activities. Subsequently, studying this type of networks as a dynamical system 
+remains rather tractable [@Moon2022]. The potential applications are numerous. Taking into account network dynamics should improve tools originally developed using non Boolean networks for the identification of interesting 
+biomarkers [@Nicolle2015] or drug repurposing. 
 
 However, the construction and analysis of Boolean networks become extremely tedious and time-consuming in the absence of experimental data or 
-when considering the activity of a large number of genes at a time [@Collombet2017]. 
-
-Moreover, the identification of interesting drug targets, via the detection of master regulators --genes at the top of the regulatory hierarchy-- suffers 
-from, first, not exploiting the full network topology, and thus, being oblivious to transcriptional regulatory cascades, which might account 
-for toxic unexpected side effects [@Bolouri2003; @Huang2019]. Second, those detection methods might not take into account the gene activity information relative to diseased 
-patients. 
-
[…]
