+++ docs/paper.rst
@@ -1,0 +1,65 @@
+Introduction to NORDic
+----------------------
+
+Genes, proteins and messenger RNAs are shown to interact on each other in order to modulate gene activity. Conversely, gene activity impacts protein production, and consequently triggers the chemical reactions needed for survival in healthy individuals. As such, perturbations of these gene regulatory interactions, through (epi)genetic and/or environmental factors, might cause diseases: e.g., the suppression of the activity of gene SCN1A is linked to a specific type of epilepsy called Dravet syndrome, both in mice and humans `[Kalume et al., 2013] <https://doi.org/10.1172/JCI66220>`_. Gene regulatory networks, which are graphs connecting biological entities according to their known regulatory interactions, are useful models that enable a better understanding of those regulatory mechanisms `[Karlebach and Shamir, 2008] <https://doi.org/10.1038/nrm2503>`_. 
+
+In particular, one type of gene regulatory networks, called Boolean networks, allows the definition of so-called regulatory functions `[Thomas, 1973] <https://doi.org/10.1016/0022-5193(74)90172-6>`_; `[Kauffman, 1969] <https://doi.org/10.1016/0022-5193(69)90015-0>`_. Those functions are specific to each node in the graph, and determine the activity of this node according to its regulators. Those functions are defined on the Boolean domain (**True** or **False**), meaning that we only consider binary gene activities. Subsequently, studying this type of networks as a dynamical system (and determining their basins of attraction, for instance) remains rather tractable `[Moon, Lee and Pauleve, 2022] <https://doi.org/10.48550/arXiv.2212.12756>`_. The potential applications are numerous. Taking into account network dynamics should improve tools originally developed using non Boolean networks, such as the Transcription Factor Influence score in CoRegNet `[Nicolle, Radvanyi and Elati, 2015] <https://doi.org/10.1093/bioinformatics/btv305>`_ for the identification of interesting biomarkers, or drug repurposing. 
+
+For the latter, a library of druggable molecules can be screened for good candidates based on the paradigm of "signature reversion"; a good drug candidate should be able to "reverse" the gene activity profile 
+associated with a diseased individual `[Duan et al., 2016] <https://doi.org/10.1038/npjsba.2016.15>`_; `[Delahaye et al., 2016] <https://doi.org/10.1186/s13059-016-1097-7>`_; `[Musa et al., 2018] <https://doi.org/10.1093/bib/bbw112>`_. That is, such a drug would stimulate abnormally weakly activated genes with respect to healthy individuals, and vice-versa. This screening approach may considerably speed up drug development, especially for rare or tropical neglected diseases `[Walker et al., 2021] <https://doi.org/10.1093/cid/ciab350>`_.
+
+However, the construction and analysis of Boolean networks become extremely tedious and time-consuming in the absence of experimental data or when considering the activity of a large number of genes at a time `[Collombet et al., 2017] <https://doi.org/10.1073/pnas.1610622114>`_.
+
+Moreover, the identification of interesting drug targets, via the detection of master regulators --genes at the top of the regulatory hierarchy-- suffers from, first, not exploiting the full network topology, and thus, being oblivious to transcriptional regulatory cascades, which might account for toxic unexpected side effects `[Bolouri et al., 2003] <https://doi.org/10.1073/pnas.1533293100>`_; `[Huang et al., 2019] <https://doi.org/10.1038/s41598-019-54180-4>`_. Second, those detection methods might not take into account the gene activity information relative to diseased patients. 
+
[…]
