+++ paper/paper.bib
@@ -1,15 +1,3 @@
-@article{Kalume2013,
-  title={Sudden unexpected death in a mouse model of Dravet syndrome},
-  author={Kalume, Franck and Westenbroek, Ruth E and Cheah, Christine S and Frank, H Yu and Oakley, John C and Scheuer, Todd and Catterall, William A and others},
-  journal={The Journal of clinical investigation},
-  volume={123},
-  number={4},
-  pages={1798--1808},
-  year={2013},
-  doi={10.1172/JCI66220},
-  publisher={Am Soc Clin Investig}
-}
-
 @article{Karlebach2008,
   title={Modelling and analysis of gene regulatory networks},
   author={Karlebach, Guy and Shamir, Ron},
@@ -78,30 +66,6 @@
   publisher={Nature Publishing Group UK London}
 }
 
-@article{Duan2016,
-  title={L1000CDS2: LINCS L1000 characteristic direction signatures search engine},
-  author={Duan, Qiaonan and Reid, St Patrick and Clark, Neil R and Wang, Zichen and Fernandez, Nicolas F and Rouillard, Andrew D and Readhead, Ben and Tritsch, Sarah R and Hodos, Rachel and Hafner, Marc and others},
-  journal={NPJ systems biology and applications},
-  volume={2},
-  number={1},
-  pages={1--12},
-  year={2016},
-  doi={10.1038/npjsba.2016.15},
-  publisher={Nature Publishing Group}
-}
-
-@article{Walker2021,
-  title={Supporting drug development for neglected tropical diseases using mathematical modeling},
-  author={Walker, Martin and Hamley, Jonathan ID and Milton, Philip and Monnot, Fr{\'e}d{\'e}ric and Kinrade, Sally and Specht, Sabine and Pedrique, B{\'e}len and Bas{\'a}{\~n}ez, Maria-Gloria},
-  journal={Clinical Infectious Diseases},
-  volume={73},
-  number={6},
-  pages={e1391--e1396},
-  year={2021},
-  doi={10.1093/cid/ciab350},
-  publisher={Oxford University Press US}
-}
-
 @article{Collombet2017,
   title={Logical modeling of lymphoid and myeloid cell specification and transdifferentiation},
   author={Collombet, Samuel and van Oevelen, Chris and Sardina Ortega, Jose Luis and Abou-Jaoud{\'e}, Wassim and Di Stefano, Bruno and Thomas-Chollier, Morgane and Graf, Thomas and Thieffry, Denis},
@@ -165,63 +129,6 @@
   organization={Springer}
 }
 
-@article{Delahaye2016,
-  title={Rare and common epilepsies converge on a shared gene regulatory network providing opportunities for novel antiepileptic drug discovery},
-  author={Delahaye-Duriez, Andree and Srivastava, Prashant and Shkura, Kirill and Langley, Sarah R and Laaniste, Liisi and Moreno-Moral, Aida and Danis, B{\'e}n{\'e}dicte and Mazzuferi, Manuela and Foerch, Patrik and Gazina, Elena V and others},
-  journal={Genome biology},
-  volume={17},
-  number={1},
-  pages={1--18},
-  year={2016},
-  doi={10.1186/s13059-016-1097-7},
-  publisher={BioMed Central}
-}
-
-@inproceedings{Margolin2006,
-  title={ARACNE: an algorithm for the reconstruction of gene regulatory networks in a mammalian cellular context},
-  author={Margolin, Adam A and Nemenman, Ilya and Basso, Katia and Wiggins, Chris and Stolovitzky, Gustavo and Favera, Riccardo Dalla and Califano, Andrea},
-  booktitle={BMC bioinformatics},
-  volume={7},
[…]
