+++ paper/paper.bib
@@ -54,18 +54,6 @@
   publisher={Nature Publishing Group UK London}
 }
 
-@article{Srivastava2018,
-  title={A systems-level framework for drug discovery identifies Csf1R as an anti-epileptic drug target},
-  author={Srivastava, Prashant K and van Eyll, Jonathan and Godard, Patrice and Mazzuferi, Manuela and Delahaye-Duriez, Andree and Van Steenwinckel, Juliette and Gressens, Pierre and Danis, Benedicte and Vandenplas, Catherine and Foerch, Patrik and others},
-  journal={Nature communications},
-  volume={9},
-  number={1},
-  pages={3561},
-  year={2018},
-  doi={10.1038/s41467-018-06008-4},
-  publisher={Nature Publishing Group UK London}
-}
-
 @article{Collombet2017,
   title={Logical modeling of lymphoid and myeloid cell specification and transdifferentiation},
   author={Collombet, Samuel and van Oevelen, Chris and Sardina Ortega, Jose Luis and Abou-Jaoud{\'e}, Wassim and Di Stefano, Bruno and Thomas-Chollier, Morgane and Graf, Thomas and Thieffry, Denis},
