### 2023-10-03 #5532 @clreda
Dear @lpantano please find in the previous message the reduced version of the paper, which I believe is less than 1,500 words now. I have removed the references which are no longer present in the paper, and posted the full version of the paper in the documentation: https://clreda.github.io/NORDic/paper.html

Let me know if that solves your concerns.
