### 2023-10-04 #5532 @Kevin-Mattheus-Moerman
@clreda I am the AEiC for this track and here to help with final steps. I have checked this review, your repository, the paper, and the archive link. Most seems in order. I only have the below points that need your attention.

On the archive:
- [ ] Please set the version tag for the ZENODO archive to `v2.4.4`, i.e. include the `v` as well. You can manually edit this.

On the paper:
- [ ] Please include the country for all affiliations.
